minor improvements to some text and formatting (#280)
QUESTION: Is there a remote participation option? If so, can you mention
in this document how to find out how to join? (Don't put the Zoom URL on
this public page, of course.)

Resolves [#ISSUE, #ISSUE]

- [ ] `docs/` have been added/updated if necessary
- [ ] `make test` has been run locally
- [ ] tests have been added/updated (if applicable)
- [ ]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated.

If you are proposing a change to the SSSOM metadata model, you must 

- [ ] provide a full, working and valid example in `examples/`
- [ ] provide a link to the related GitHub issue in the `see_also` field
of the linkml model
- [ ] provide a link to a valid example in the `see_also` field of the
linkml model


[Add a description, mentioning at least relevant #ISSUE and how it was
addressed. A bulleted list of all changes performed by the PR is is
helpful.]