Do not allow `curie_map` to be a link to a curie map.
Remove the bit in the specification that suggested the `curie_map`
parameter in the YAML metadata of a mapping set could be a link to an
external curie map. `curie_map` should only be used to provide an
embedded curie map. It is unclear whether an external curie map should
be allowed at all, but if it is, it should be specified using a distinct
parameter.