Fix SSSOM Schema not having bijective prefix map (#302)
There are issues in https://github.com/mapping-commons/sssom-py/pull/397
because the base schema does not have a valid prefix map - `dc` and
`dcterms` are both used as prefixes for the same URI prefix. This
happens because LinkML makes some inference about what needs to be there
- there are some assorted uses throughout the schema definition with
both DC and DCTERMS. This PR updates the schema to only use DCTERMS and
puts explicit entries for DCTERMS in the prefix map to address this. It
then regenerates the whole project.