Change sempav to semapv in the tutorial page (#317)
Hi mapping commons, here with a suggestion for the tutorial page. 

Googled to see which is the recommended prefix but I'm still not sure to
which one you will stick. The different spelling confused me a little
bit when following the tutorial, is this something you would like to
address?