Recommend that YAML keys be alphabetically sorted.
This commit adds a recommendation that the keys in the YAML metadata
block (in both "embedded" and "external" modes) SHOULD be sorted by
case-sensitive alphabetical order.

There is no particular reason for recommending alphabetical order over
any other type of order, beyond the fact that sssom-py is already using
alphabetical order (probably because that is seemingly the default
behaviour of the PyYAML's 'safe_dump' method used by sssom-py). So we
just take the order that is de facto in use in the wild and make it the
officially recommended order.