upgrade `mkdocs-mermaid2-plugin` major version (#353)
This PR upgrades the major version of the `mkdocs-mermaid2-plugin`
package which is asserted as a dependency on this library. The reason
why this is necessary is because
[oaklib](https://github.com/INCATools/ontology-access-kit/) depends on
`sssom-schema` and other downstream packages which use `oaklib` as a
dependency are not able to enforce the latest versions of
`mkdocs-mermaid2-plugin` in their repos because of conflicting
dependency versions.

- [ ] `docs/` have been added/updated if necessary
- [x] `make test` has been run locally
- [ ] tests have been added/updated (if applicable)
- [ ]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated.