Replace index page.
Use the 'home.md' file, renamed to 'index.md', as the "index page" (that
is, the page shown when the visitor does not request an explicit page),
instead of the LinkML-generated index page (which is not suitable to be
the first thing the visitor sees upon arriving on the website).

The LinkML-generated index page is renamed to 'linkml-index.md'. The
name does not really matter, as long as it is *not* 'index.md'.

This requires updating LinkML to version 1.7.0 at least, because prior
versions hardcoded the name of the generated index page to 'index.md',
thereby forcing that page to be the site index page (unless the web
server is configured differently). This in turns requires bumping
slightly the minimum Python version from 3.8 to 3.8.1.