Fix examples of SSSOM/TSV files.
The examples provided in the SSSOM/TSV section of the "overview"
document are full of errors and would fail the most basic validation by
our own tools:

- use of "Lexical" instead of "semapv:LexicalMatching" in the
  mapping_justification field (probably a remnant of the time prior to
  the adoption of the SEMAPV vocabulary);
- bogus IRI prefix for the SKOS namespace (missing terminal '#');
- use of a full-length identifier (instead of a CURIE) for `creator_id`.

This commit fixes those errors. In addition, it also ensures that the
fields are listed in the *recommended order*. It's not critical but if
we take the time to recommend that fields be sorted in a given order,
the least we can do is to follow our own advice in our examples.

While we are at it, we also add a small note about the requirement for
using CURIEs in the SSSOM/TSV format, since that requirement currently
does not appear anywhere but is already enforced by `sssom validate`.