Update documentation for entity reference to clarify CURIE/URI type (#358)
Resolves https://github.com/mapping-commons/sssom-py/issues/512

- [X] `docs/` have been added/updated if necessary

The documentation for "entity references" is largely confusing, and
leading, in its current form, to the expectation that metadata fields
types with these can be _either_ a CURIE _or_ a full URI.

This is not quite right: in the TSV serialistion, as well as JSON, we
expect the values to be CURIE-strings. This PR documents this design
decision, but, as @joeflack4 puts it rightfully, this has a bit of a
"smell" to it. Not sure what the right thing is here, but this
documentation here is certainly better than not having it.