Create placeholder documents for the actual spec.
Split the existing "spec.md" file into two components:

- a general introduction on mappings;
- the actual specification, which is itself split in several parts:
    - the specification of the data model;
    - the specification of the serialisation formats.

This commit creates placeholder files that will hold those different
sections. The "general introduction" file is pre-filled with the
contents of the "Introduction" section of the original "spec.md" file.

(I believe that introduction should be entirely re-written from scratch,
as it sometimes reads like a patchwork of unrelated pieces pasted
together. But that will be for later work. The most urgent for now is to
have a place where we can write the actual *specification* with all its
details.)