Add list of common mapping predicates.
In the section about the data model, we add a list of the mapping
predicates that are considered common and that are recommended.

This is mostly taken from the old spec.md document, except that we also
mention the predicates defined in the SEMAPV vocabulary, that the old
spec was not mentioning at all.