Add specifications for the serialisation formats.
Add a complete and workable specification for the SSSOM/TSV format. This
is an original version that takes very little from the old spec.md
document, except the examples.

The specification for the OWL/RDF format, on the other hand, is directly
taken from the old spec, almost "as is".

The "specification" for the JSON format is currently merely a
placeholder, since that format is NOT specified for now.