Fix canonical format for floating values.
Amend the canonical rule for serialising floating point values with UP
TO 3 digits after the decimal point AS NEEDED.

That is, if more than 3 digits would be needed to write the value, then
the writer MUST truncate after the third digit, but if the value can be
written (without loss of precision) with less than 3 digits, the writer
MUST NOT right-pad the value with zeroes.

So a value like 0.9 is to be written as "0.9", NOT as "0.900".