Reorganise docs (#368)
Resolves [#330]

- [x] `docs/` have been added/updated if necessary
- [x] `make test` has been run locally
- [ ] tests have been added/updated (not applicable)
- [ ]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated (not applicable)

This PR reorganises the documentation, especially the specification
part, as suggested in #330.

More precisely:

* Several bits of informations that were scattered throughout the
website are now regrouped on the index page (e.g. “contacts” and
“credits”).
* The home page gets a “SSSOM at a glance” section to give an immediate
glimpse of what the standard is about.
* The specification is broken down as follows:
  * general introduction
  * specification of the data model
* introduction and notes complementary to the LinkML-generated
documentation
    * LinkML-generated documentation
  * specification of the serialisation formats
    * SSSOM/TSV format (completely rewritten)
    * OWL/RDF format (taken from the pre-existing `spec.md` document)
* JSON format (currently merely a placeholder, since the JSON format is
unspecified #321)

The “resources for users” section is left untouched for now. The urgent
part was reorganising the _specification_, so that we can start
enriching it to make it ready for 1.0.