Mark propagatable slots as such in the LinkML model.
This commit adds a "propagated" annotation to all slots that should be
considered "propagatable", that is, all slots that, when set at the
level of the mapping set, can (and should) be propagatated down to the
level of individual mappings.

The presence of the annotation in the model replaces the list available
in the src/docs/spec-model.md document as the authoritative source for
which slots are propatagable or not. The aforementioned list remains in
place, but only as a convenience for readers of the spec.