Add min/max value constraints to score and confidence slots. (#373)
Resolves #372

- [ ] `docs/` have been added/updated if necessary (not applicable; the
doc was already ahead of the model, this PR is about updating the model
to align it with the doc)
- [x] `make test` has been run locally
- [ ] tests have been added/updated (not applicable)
- [x]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated.

The description of the `confidence`, `semantic_similarity_score`, and
`similarity_score` slots says that those slots expect a value between 0
and 1, but the model does not enforce that. LinkML provides a way to
formalise this kind of value constraints (which could then be
automatically enforced by a runtime that is aware of those constraints),
so we do exactly that here: we add `minimum_value` and `maximum_value`
constraints to the aforementioned slots.