Document testing SSSOM-Py against a development version of the schema.
Update the CONTRIBUTING.md document to add a section about how to
locally run the test suite, and especially how to locally run SSSOM-Py's
test suite.

Update the pull request template to add a checkbox to remind
contributors that any change to the LinkML schema should be tested
against SSSOM-Py (or rather, SSSOM-Py should be tested against the new
schema).