Add "extension definition" to the model.
We add a new slot to the MappingSet object , representing the
definitions of extension slots that may be used within the set (as a
list of "extension definition" objects).

Each "extension definition" is made of up to three elements:

* the name of the extension slot,
* the identifier of an associated property,
* a type hint.

We also add a sample file that demonstrates the use of such extension
slots.