Misc fixes.
Clarify reference to `slot_name` in the description of the `property`
slot.

Fix incorrect reference to a SSSOM/TSV _writer_ (not _reader_) when we
state that unknown slots MUST NOT be written in SSSOM/TSV files.

Replace "catenate" by "concatenate". My understanding is that the two
terms are actually equivalent, but "concatenate" is for some reason more
widely used at least in computer science.