Add new "literal profile" solution based on "subject_label" and "object_label" (#384)
Resolves #234

This PR introduces an alternative model to the "literal mapping"
proposal we have previously added. It is built on the following
principles:

- Introducing a profile causes a needless overhead which requires
specialised tooling
- We can achieve the same functionality with a _very_ minimal
intervention in the main metadata model

I do understand that there are various opposing views on the need for a "literal" profile, but I think this super minimal intervention will satisfy both sides. In essence, we do not have a literal profile; we have a convention that allows us to represent an "entity" by its label (subject_label) rather than by a semantic identifier. This means we do not need specialised tooling and documentation (nor training).

Huge thanks to @gouttegd 🙏 who managed to steer this massive carrier ship after it had left the harbor. This is rarely successful and needed a huge amount of thought, testing, and patience (mostly with me and my constant questions), and I am supremely happy we managed to make it!

---------

Co-authored-by: Damien Goutte-Gattat <dpg44@cam.ac.uk>
Co-authored-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>