Update the landing page, and list of contributors, steering committee
The main update here is to the list of contributors, and the formal definition separation of editor and steering committee. I do understand that concept of "steering committee" has never been formally discussed - as co-founder and main developer I am taking the liberty to define this group, that has effectively operated but not named so far, now.