Add the `predicate_type` slot to Mapping and MappingSet classes. (#405)
Resolves [#404]

- [x] `docs/` have been added/updated if necessary
- [x] `make test` has been run locally
- [ ] ~~tests have been added/updated (if applicable)~~
- [x]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated.

If you are proposing a change to the SSSOM metadata model, you must 

- [x] provide a full, working and valid example in `examples/`
- [x] provide a link to the related GitHub issue in the `see_also` field
of the linkml model
- [x] provide a link to a valid example in the `see_also` field of the
linkml model
- [x] run SSSOM-Py test suite against the updated model

The `predicate_type` slot has been _defined_ but never _used_ in any
classes. Presumably this is an oversight, so we add it here to the
Mapping and MappingSet classes. We make it a “propagatable” slot, for
consistency with the `subject_type` and `object_type` slots.