Allow using `similarity_measure` at the set level. (#412)
We add the `similarity_measure` slot to the MappingSet class, as a
propagatable slot. It should be quite common that all `similarity_score`
for all mappings in a set are determined using the same measure, so it
makes sense to allow setting the measure once and for all at the set
level, without having to repeat that information for all individual
mappings.

closes #411