Change type of `see_also` to `xsd:anyURI`. (#424)
Resolves [#422]

- [ ] `docs/` have been added/updated if necessary
- [x] `make test` has been run locally
- [ ] tests have been added/updated (if applicable)
- [x]
[CHANGELOG.md](https://github.com/mapping-commons/sssom/blob/master/CHANGELOG.md)
has been updated.

If you are proposing a change to the SSSOM metadata model, you must 

- [ ] provide a full, working and valid example in `examples/`
- [x] provide a link to the related GitHub issue in the `see_also` field
of the linkml model
- [ ] provide a link to a valid example in the `see_also` field of the
linkml model
- [x] run SSSOM-Py test suite against the updated model

The `see_also` slot is explicitly described as expecting URL values, so
there should not be any reason for it to be typed as `xsd:string`. We
re-type it as a `xsd:anyURI` instead.