Add the sssom_version slot.
We add a new set-level slot to indicate the version that a set is
compliant with.

We also extend the metamodel with a new extension class ("Versionable")
that we use to explicitly mark a slot (such as the new `sssom_version`
slot) as having been introduced in a specific version of SSSOM.

Both the `sssom_version` slot and the `Versionable#added_in` annotation
uses a new enum, `sssom_version_enum`, to represent the version of the
SSSOM specification. For now, that enum contains only two allowed
values, representing the current published version of the spec (1.0) and
the upcoming version (1.1).

We update the documentation of the model to explain the versioning
scheme we use, what are our promises regarding backwards compatibility,
and how to deal with the new `sssom_version` slot.