Mark `predicate_type` as having been added in 1.1.
Though the `predicate_type` slot had been in the schema since (AFAIK)
the beginning, it has only been added to the Mapping and MappingSet
classes *after* the publication of version 1.0 of the spec. That slot
should therefore be considered a novelty of the upcoming version 1.1,
since it was *not* usable in version 1.0.