Update contributing guidelines for backwards compatibility.
Add a section to the CONTRIBUTING.md file to remind would-be
contributors of the need to consider backwards compatibility, and to
show how new slots should be tagged with the version of the spec they
are introduced with.

Also add a checkbox to the pull request template to remind contributors
that new slots should be tagged as such.