Update contributing guidelines to mention backwards compatibilty (#447)
This PR updates the contributing guidelines to explain that changes to
the model in the 1.x branch should be ideally limited to adding optional
slots, and that added slots should be tagged with the `added_in`
annotation to indicate the version of the spec they will be introduced
in.

Arguably this could/should have been done as part of the PR in which the
`sssom_version` slot was introduced (#440).