Clarify the absence of link between author_id and author_label.
There is not supposed to be any link between `author_id` and
`author_label`. In particular, `author_label[n]` is _not_ intended to be
the human-friendly label of `author_id[n]`.

This has seemingly always been the intention (see notably this comment:
https://github.com/mapping-commons/sssom/issues/344#issuecomment-1893588484)
but has never been made explicit in the schema, which is what we do
here.

Likewise for reviewer_id/reviewer_label and creator_id/creator_label.