Add `record_id` slot.
Add a new slot to the Mapping class, `record_id`, intended to hold a
unique identifier for a given mapping.

The slot is optional, so as not to break compatibility with existing
SSSOM 1.0 sets. This also means that, while we can define the slot as a
"unique key" for the Mapping class, we _cannot_ define it as the
"identifier", because in LinkML identifier slots are automatically
mandatory.

The identifier is intended to be completely opaque. How to generate
identifiers is left to the producers of SSSOM sets, and no meaning of
any sort should be assigned to an identifier.

closes #359