Clarify that URI-typed slots expect absolute URI values.
Add a new type `AbsoluteURI`, based on the LinkML type `uri` but with
the added restriction that the URI must be absolute. Update `uri`-typed
slots to make them use the `AbsoluteURI` type instead. Add a
recommendation that implementations SHOULD still accept relative URIs in
all `AbsoluteURI`-typed slots, when they are processing a SSSOM 1.0 set.

closes #448