Revise definition of sssom:AbsoluteURI (renamed to sssom:NonRelativeURI).
We rename the `sssom:AbsoluteURI` to `sssom:NonRelativeURI`. "Absolute
URI" is a misnomer and a possible source of confusion, because the RFC
3986 uses that name to refer to URIs that both (1) have a scheme
component and (2) do not have a fragment component. In the context of
SSSOM, we only want (1): we want to forbid relative URI references that
do not have a scheme, but we do _not_ want to forbid non-relative URIs
that have a fragment.

(The name `NonRelativeURI` is somewhat ridiculous, as per the RFC 3986
all URIs are necessarily non-relative, and only "URI references" can be
relative; but we cannot name our type "URI" because of the risk of
confusion with the LinkML type, which despite its name refers in fact to
URI references... ><)

We also explicitly clarify that the type is specifically intended to
exclude relative URI references, and provide some examples of both valid
URIs and invalid relative URI references.