Assign a property to ALL enum values. (#458)
Related to https://github.com/mapping-commons/sssom/discussions/454

This PR ensures that all enum values (_except_ for the
`mapping_cardinality_enum`) throughout the LinkML model have a property
assigned to them. This is important for RDF serialisation, so that all
enum values are consistently rendered as resource identifiers, instead
of having some values being rendered as resource identifiers and some
values being rendered as string literals.

I normally agree with @nichtich ’s sentiment (expressed in #450) that we
should as much as possible reuse existing properties from outside SSSOM
(as we already do, for example, for most values of the
`entity_type_enum`) and avoid creating SSSOM-specific properties in the
`https://w3id.org/sssom/` namespace. However, for all the enum values
concerned in this PR, I am not aware of pre-existing properties that
would be suitable, so I did in fact create a bunch of SSSOM-specific
properties. I am open to replacing some of them by more appropriate
existing properties if someone can suggest some.

Also open to discussing the naming of those properties.

Of note, it would be nice if those properties could resolve to something
meaningful to users. In particular, for the values in
`sssom_version_enum`, I think most people would reasonably expect
`https://w3id.org/sssom/version1.0` to resolve to a frozen version of
the full specification as it was in the 1.0 release (and likewise for
`https://w3id.org/sssom/version1.1` and any futur version). But this is
an infrastructure issue that I consider out of scope for this PR.

---------

Co-authored-by: Emily Hartley <ehartley@c-path.org>