Clarify that URI-typed slots expect non-relative URI values. (#453)
Resolves [#448]

URI-typed slots are expected to contain (non-relative) URIs only, but
the LinkML `uri` type accepts both URIs _and_ relative URI references
(something that is not very clear in the LinkML’s metamodel, but [the
implementation of the `URI`
type](https://github.com/linkml/linkml/issues/2726#issuecomment-2909331813)
leaves no doubt that relative URI references are considered valid
values).

So we add a new `NonRelativeURI` type, based on the LinkML `uri` type
but with the added restriction that the URI must not be a relative URI
reference. Of note, that restriction is only _documented_, not
_enforced_: it is up to implementations to actually check that the value
of a `NonRelativeURI`-typed slot is indeed a non-relative URI, as the
underlying LinkML implementation will accept both non-relative and
relative URI references.

All slots that were previously defined with a `uri` range are redefined
to use the new `NonAbsoluteURI` type.

Since it has never been clear, in SSSOM 1.0, that URI-typed slots were
supposed to contain only non-relative URIs, for backwards compatibility
it is RECOMMENDED that implementations still accept relative URI
references when processing sets compliant with version 1.0 of the spec.

---------

Co-authored-by: Nico Matentzoglu <nicolas.matentzoglu@gmail.com>
Co-authored-by: Emily Hartley <ehartley@c-path.org>