Fix chomping of description blocks for subject|object_match_field.
The description of the `subject_match_field` slot is written as a folded
scalar block (`> ...`) with no exlicit chomping indicator. As a result,
the last newline character is interpreted as being part of the
description.

This in turn messes with the formatting of the slot table on SSSOM's
index page, because the row for the `subject_match_field` slot is spread
over two lines (since the description cell contains a newline
character):

  | subject_match_field | description...
  entity reference | rdfs:label |

instead of the expected:

  | subject_match_field | description... | entity reference | rdfs:label |

This commit adds an explicit chomping indicator (`>+`) to prevent the
terminating newline character from being taken as part of the scalar
value.

Likewise for `object_match_field`.