Add `cardinality_scope` slot. (#476)
Resolves [#467]

Adds a new mapping record slot called `cardinality_scope`,
intended to precise the meaning of the value of the
`mapping_cardinality` slot.

If empty, it means that the cardinality is relative to the entire set
(which corresponds to the current definition of `mapping_cardinality`,
and therefore ensures backwards compatibility). If non-empty, it is
expected to be a list of slots that are used to define the subset (the
"scope") of mapping records in which the cardinality has been computed.

For example, a cardinality scope of (`predicate_id`, `object_source`)
means that the cardinality is computed over the subset of mapping
records that have the same predicate and whose objects come from the
same source.

The new slot is propagatable, because it is assumed that in most mapping
sets, the cardinality for all records will be computed in the same way
(though there's no obligation for that), and so all records will share
the same scope.