Annotate class-specific usage of slot with `added_in`.
The `added_in` annotation is used to indicate that a slot has been
introduced in a specific version of the specification. For now, that
annotation is only used in the model-wide definition of slots, which
means it can only indicate when a brand new slot has been introduced --
it can not be used to indicate that a pre-existing slot has been added
to a class in which it was not used before.

For example, the `similarity_measure` slot has existed since before 1.0,
but has been only been added to the MappingSet class for version 1.1.

In this commit, we make use of LinkML's `slot_usage` mechanism to add
the `added_in` annotation not only to the general (model-wide)
definition of a slot, but also to the specific instance of a slot within
a class.