Add 0:0 special cardinality value.
This commit adds a new cardinality value of `0:0`, to be used for mapping
records where both the `subject_id` and the `object_id` are
sssom:NoTermFound. Such records are allowed and may be used to represent
the fact that there is no overlap between a subject vocabulary and an
object vocabulary.

We also try to explain a bit more what the different cardinality values
mean exactly, and provide two examples:

* one that illustrates all possible cardinality values between real
  mappings (1:1, 1:n, n:1, n:n);
* one that illustrates the special cases where the subject_id and/or the
  object_id is sssom:NoTermFound.

closes #477