Add test for added_in annotation.
We add a test that any new slot must

(1) carry a `added_in` annotation that
(2) is set to the current version of the spec.

The test works by preparing a fixed compiled list of all the "non-new"
slots (all slots present in the 1.0 release) for both the Mapping and
MappingSet classes. It then loads the current schema and iterates over
all slots for the Mapping and MappingSet classes, checking that every
slot that is not a pre-existing slot is annotated as expected.