Add test for added_in annotation. (#483)
This PR adds a test that any new slot must

(1) carry a `added_in` annotation that
(2) is set to the current version of the spec.

The test works by preparing a fixed compiled list of all the “non-new”
slots (all slots present in the 1.0 release) for both the `Mapping` and
`MappingSet` classes. It then loads the current schema and iterates over
all slots for those two classes, checking that every slot that is not a
pre-existing slot is annotated as expected.

Of note, the test currently fails (as it should), because the
`similarity_measure`, `curation_rule`, and `curation_rule_text` slots
are _new_ on the `MappingSet` class but are currently not annotated as
such (this is what #482 is fixing).