Move definition of propagation/condensation outside of SSSOM/TSV.
Propagation and condensation of "propagatable" slots is currently
defined in the specification for the SSSOM/TSV format. This is not
appropriate as those operations are completely independent of the
serialisation format -- propagation and condensation are performed
identically regardless of whether a mapping set is read/written from/to
a SSSOM/TSV file, a SSSOM/JSON file, or a RDF file.

So here, we move the definition of those operations to the general
introduction (serialisation-independent) to the SSSOM model.

We do not change anything about how propagation and condensation are
performed, and the associated requirements.

closes #486