Fix bogus indentation in RDF spec (#498)
This PR fixes a small formatting issue in the recently merged RDF
specification.

The Markdown implementation used by Mkdocs only recognizes nested lists
if the nested list items are indented by precisely *4 spaces* (or a tab
character) compared to the parent list item.

This was not caught prior to merging because I (and likely anyone who
reviewed the PR) relied on GitHub’s Markdown preview, but the
Github-flavoured Markdown is slightly different from the Mkdocs
Markdown. Therefore the PR also amends the contributing documentation to
recommend that documentation editors run `make serve` locally.